Create Diamonds
You are given a sequence of n members of the numbers 0 and 1 whose members are numbered from 1 to n. In each operation, you can turn a subset of this sequence upside down. The cost of each operation is equal to the length of the selected subroutine. We ask you to arrange the given sequence in descending order by performing a number of the mentioned operations so that the total cost of the operations does not exceed 20 × n.