Create Divide and conquer
We have a playground n × n whose rows are numbered from 1 to n from top to bottom and its columns are numbered from 1 to n from left to right. In m there is a house of the diamond map. Our nut (extraction machine) is a 1 × L rectangle (note that L can also be 1) and we first place it horizontally in the desired position on the first row (top). At each stage of the game, our piece goes down one row and can move up to one unit horizontally (it can also not move horizontally). In each position of the game, our bead gets all the diamonds on it. We want to find the minimum value of L so that our bead can get all the diamonds.